requirements.txt
numpy>=1.21.0
torch>=1.9.0
transformers>=4.5.0
datasets>=1.8.0
nltk>=3.6.0
scikit-learn>=0.24.0
pyspellchecker>=0.6.0
autocorrect>=2.6.0
textblob>=0.15.3
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
spello>=1.3.0
editdistance>=0.8.1
pytest>=6.2.5
pytest-benchmark>=3.4.1
sentencepiece>=0.1.99 